Fix Stupid BLog error and smarter leaving warning
Fixed an error caused by the BattleScreen loop that is trying to update
an element that no longer exists.
Made the 'your leaving the page while your in a queue' warning to be
smarter. It will only show if you are in a queue where leaving would
cancel it but wont show if you're already in a game where it can be
safely closed.